refactor: split up system-usage component
